[Refactor] Split setup page flows (#785)
